Fix panic when cache file can't be Stat-ed
The code currently assumes that Stat always return a FileInfo object,
but if the file can't be Stat-ed (e.g. due to a permissions error), Stat
can return a nil FileInfo, plus an error. The error is only handled if
it's a "not exists" error, but there are other errors that can be
returned too.

This wasn't caught by the tests, since the mocked out Stat always
returns a FileInfo, even when an error is returned.

This commit also changes from os.IsNotExist(err) to errors.Is(err,
fs.ErrNotExist) as recommended by the documentation for os.IsNotExist.