Add support to optionally exclude each subcommand
The current default is to compile all functionality into the
aws-iam-authenticator binary, however the tool is often used as
exclusively client-only (on nodes / local machines) or exclusively
server-only (ie in the EKS control plane).

We can dramatically reduce unneeded binary size by adding the option to
exclude subcommand which are known to be unneeeded at build time.

```
> make bin && ls -la _output/bin
-rwxrwxr-x 1 emily emily 53559296 May 13 13:54 _output/bin/aws-iam-authenticator

> make bin GOTAGS="no_init no_verify no_add no_server"
-rwxrwxr-x 1 emily emily 18907136 May 13 13:54 _output/bin/aws-iam-authenticator
```