Keep original aws-auth configuration when the new value cannot be parsed
In the current implementation, a value like `rokeMappings` is always
reinitialized to an empty slice for any `Added` or `Modified` event:

`roleMappings = make([]config.RoleMapping, 0)`

And then, even when there's parse errors detected, it's passed to
`saveMap` with other probably valid data, and inside `saveMap` the field
`ms.roles` is, again, always reinitialized to an empty slice.

We have introduced an error in the `mapRoles` configuration recently
like the following:

`time="2023-06-13T09:35:43Z" level=warning **msg="Errors parsing
configmap: [yaml: line 13: did not find expected key]"**`

And this change alone broke the whole EKS cluster because, as explained
above, `ms.roles` is always reinitialized to an empty slice and now even
the nodes in cluster get `access denied` because of the `ARN is not
mapped` error.

*To make EKS clusters less fragile*, we should keep the original
configuration if the new one is malformed.